+++ src/main/java/org/http/support/BaseHttpAcceptor.java
@@ -6,24 +6,17 @@ import org.http.HttpRequest;
 import org.http.chain.HttpHandler;
 import org.http.chain.HttpSession;
 import org.http.chain.support.BaseHttpService;
-import org.http.chain.util.ExceptionMonitor;
 import org.http.exception.HttpInvokeException;
 import org.http.exception.HttpSessionClosedException;
 import org.http.executor.GlobalHttpExecutor;
 import org.http.executor.HttpExecutor;
 
 public abstract class BaseHttpAcceptor<T> extends BaseHttpService implements HttpAcceptor<T> {
-
 	
 	/**
 	 * http客户端 2016年1月18日 下午6:00:01
 	 */
 	protected final HttpClientFactory httpClientFactory;
-
-	/**
-	 * 判断filterChain是否build 2016年1月20日 下午1:25:48
-	 */
-	private volatile boolean isBuildChain;
 	
 	private HttpExecutor executor;
 
@@ -47,20 +40,8 @@ public abstract class BaseHttpAcceptor<T> extends BaseHttpService implements Htt
 	public T service(HttpRequest request, HttpHandler handler) throws HttpSessionClosedException, HttpInvokeException  {
 		// 准备service
 		HttpSession session = prepareService(request, handler);
-		
-		try {
-			getFilterChainBuilder().buildFilterChain(session.getFilterChain());
-			// 没有build链则build
-//			if (!isBuild()) {
-//				
-//				getFilterChainBuilder().buildFilterChain(session.getFilterChain());
-//				build(true);
-//			}
-		} catch (Exception e) {
-			ExceptionMonitor.getInstance().exceptionCaught(e);
-		}
 
-		return doService(request, handler, session);
+		return doService(session);
 	}
 
 	/**
@@ -77,19 +58,8 @@ public abstract class BaseHttpAcceptor<T> extends BaseHttpService implements Htt
 	 * @return 2016年1月20日 下午1:45:35
 	 * @throws Exception
 	 */
-	protected abstract T doService(HttpRequest request, HttpHandler handler, HttpSession session)
+	protected abstract T doService(HttpSession session)
 			 throws HttpSessionClosedException, HttpInvokeException ;
-
-
-	@Deprecated
-	public boolean isBuild() {
-		return isBuildChain;
-	}
-
-	@Deprecated
-	public void build(boolean build) {
-		isBuildChain = build;
-	}
 
 	public HttpClientFactory getHttpClientFactory() {
 		return this.httpClientFactory;